+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - CHANGE Split conflict handler functionality into `isEqual()` and `resolve()`.
 - CHANGE (internal) remove `conflictResultionTasks()` and `resolveConflictResultionTask()` from the RxStorage interface.
 - FIX (GraphQL replication) datapath must be equivalent for pull and push [#6019](https://github.com/pubkey/rxdb/pull/6019)
+- CHANGE `ignoreDuplicate: true` on `createRxDatabase()` must only be allowed in dev-mode.
 <!-- ADD new changes here! -->
 
 <!-- /CHANGELOG NEWEST -->